+++ python/kicad_interface.py
@@ -341,6 +341,7 @@ class KiCADInterface:
             "add_net": self.routing_commands.add_net,
             "route_trace": self.routing_commands.route_trace,
             "check_route_segment": self.routing_commands.check_route_segment,
+            "find_via_lane": self.routing_commands.find_via_lane,
             "add_via": self.routing_commands.add_via,
             "delete_trace": self.routing_commands.delete_trace,
             "dedupe_traces": self.routing_commands.dedupe_traces,